add bras1 wlanacname support
Since the autumn semester of 2023, in dormitory 4 some change happened. When doing webauth, sometimes the POST message will be send to a link which parameter "wlanacname" is "web.wlan", but sometimes it will be send to a link which "wlanacname" is "bras1"(I haven't seen this before).
As I failed to find the regular pattern, that is, I don't know when it will POST the message to "web.wlan", when it will be "bras1", so I just put these two curl commands together. I guess such a change was because new friends came into our school, so those APs on higher floors was turned on, but there are some mistakes in settings.
Also in this file there was a quotation mark mismatch. Fixed it.